Depend on the new rust-psbt crate
There is a beta release of a new `psbt-v2` crate. The crate includes all
the code from `rust-bitcoin` and `rust-miniscript`'s `psbt` modules from
the 0.31.1 and 11.0.0 releases respectively.

We can use the new crate as a drop in replacement with minimal changes.

Then, when the `psbt-v2` non-beta release is out we `bdk` could consider
using it if they want.

Requires patching `rust-hwi` also.